% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_methods.R
\name{cca}
\alias{cca}
\title{Canonical Correlation Analysis}
\usage{
cca(X, Y, tol = .Machine$double.eps)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}
}
\value{
a list of results including
\itemize{
\item \code{d}: vector of singular values
\item \code{u}: matrix of left singular vectors
\item \code{v}: matrix of right singular vectors
\item \code{l}: vector of eigenvalues
\item \code{lx}: matrix of latent variables for X
\item \code{ly}: matrix of latent variables for Y
\item \code{p}: matrix of left generalized singular vectors
\item \code{q}: matrix of right generalized singular vectors
\item \code{m}: left metric matrix for the GSVD. For CCA, it is the inverse crossproduct of \code{X}.
\item \code{w}: right metric matrix for the GSVD. For CCA, it is the inverse crossproduct of \code{Y}.
\item \code{f}: matrix of loadings for \code{X}. \code{lx} is computed by multiplying \code{X} and \code{f}. For CCA,
\code{f} is \code{m} multiplied by \code{p}.
\item \code{g}: matrix of loadings for \code{Y}. \code{ly} is computed by multiplying \code{Y} and \code{g}. For CCA,
\code{g} is \code{w} multiplied by \code{q}.
}
}
\description{
A basic function for computing results of Canonical Correlation Analysis (CCA).
Relies on the underlying \code{\link[GSVD:gplssvd]{GSVD::gplssvd()}} method.
}
