% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{diff_ratio}
\alias{diff_ratio}
\title{Compute difference or ratio between consecutive values in a vector}
\usage{
diff_ratio(vec, method = "diff")
}
\arguments{
\item{vec}{an input vector of numerical values (such as eigenvalues)}

\item{method}{"diff" (default) or "ratio"}
}
\value{
a vector the same length as \code{vec}. The last value is \code{NA}.
}
\description{
This function returns a vector the same length as its input vector \code{vec}
where each position has the following position subtracted from it or is
divided by it. This can be used with a vector of eigenvalues to get the
eigenvalue differences or eigenvalue ratios.
}
