% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRP_DRP.R
\name{DRP}
\alias{DRP}
\title{Deflation randomization procedure (DRP)}
\usage{
DRP(
  X,
  Y,
  ttmethod,
  n_iter = 99,
  rand_type = "rand_rows",
  tol = .Machine$double.eps,
  parallel = FALSE,
  n_dim_comp = NULL
)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{ttmethod}{string denoting the two-table method to use,
either 'plsc', 'cca', or 'rda'.}

\item{n_iter}{(default: 99) number of permutation iterations}

\item{rand_type}{(default: 'rand_rows') how to randomize the matrices,
either 'rand_rows'/'RR' or 'rand_each_col'/'REC'.
The RR option permutes entire rows of the \code{X} matrix while leaving \code{Y} unchanged.
The REC option permutes every column of both \code{X} and \code{Y}.}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}

\item{parallel}{(default: FALSE) use a parallelized loop to do the main permutations.
This will use the \code{parallel} package to utilize multiple cores simultaneously
(using \code{\link[parallel:detectCores]{parallel::detectCores()}} - 1 cores).}

\item{n_dim_comp}{(default: \code{NULL}) the number of dimensions to compute null distributions for.
\code{NULL} means all dimensions are tested.
Testing a smaller number of components will cut down on computational time.}
}
\value{
See \code{\link[=SRP]{SRP()}} for the list of test statistics and null distributions that are returned,
with the only difference that \code{DRP()} does not include \code{l_prop} or \code{l_prop_dist}.
}
\description{
DRP is a two-table stopping rule that runs a sequential permutation test with deflation.
Each component is tested individually, and the effects of previous components are
removed using deflation.
The function computes several test statistics and their null distributions.
It does not actually compute the p values for each component; that is accomplished by
passing the results to \code{\link[=pval_from_dist]{pval_from_dist()}} or \code{\link[=pvals_all]{pvals_all()}}.
}
