% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_invsqrt_psd_matrix}
\alias{fast_invsqrt_psd_matrix}
\title{Inverse square root of positive semi-definite matrix}
\usage{
fast_invsqrt_psd_matrix(x, ...)
}
\arguments{
\item{x}{positive semi-definite matrix}

\item{...}{everything else passed to \code{tolerance_eigen}}
}
\value{
the inverse square root matrix
}
\description{
This function was stolen from the GSVD package by Derek Beaton
and modified to eliminate the input checks
and allow for additional arguments to be passed to \code{tolerance_eigen} and by extension
\code{eigen}. This makes it much faster by passing the \code{symmetric} argument when
matrix symmetry is known.
}
