% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cormats_and_data.R
\name{gen_mvrnorm_from_cor_fast}
\alias{gen_mvrnorm_from_cor_fast}
\title{Generate multivariate normal data from a correlation matrix.}
\usage{
gen_mvrnorm_from_cor_fast(cor_mat, nrows, ncolX, ncolY, mu = NULL, seed = NULL)
}
\arguments{
\item{cor_mat}{a matrix of size \code{ncolX + ncolY} by \code{ncolX + ncolY}. This is
the joint correlation matrix of \code{X} and \code{Y}, likely created using \code{\link[=multi_block_cor]{multi_block_cor()}}.}

\item{nrows}{integer number of rows to sample}

\item{ncolX}{integer number of columns for \code{X}. Rows/columns 1:ncolX of \code{cor_mat}
correspond with \code{X}.}

\item{ncolY}{integer number of columns for \code{X}. Rows/columns (ncolX+1):(ncolX+ncolY)
of \code{cor_mat} correspond with \code{Y}.}

\item{mu}{vector of means for each column. default is \code{NULL} which sets all means
to zero.}

\item{seed}{(default:NULL) optional seed for reproducibility. Because of the
underlying C++ code, the normal R \code{set.seed()} will not work and it must be
passed to the function instead.}
}
\value{
list containing:
\itemize{
\item \code{X}: matrix of size \code{nrows} by \code{ncolX}
\item \code{Y}: matrix of size \code{nrows} by \code{ncolY}
}
}
\description{
This function samples two tables of data, \code{X} and \code{Y},
from a multivariate normal distribution using the joint correlation matrix
between \code{X} and \code{Y}, likely created using \code{\link[=multi_block_cor]{multi_block_cor()}}.
}
