% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_POV.R
\name{kaiser_guttman}
\alias{kaiser_guttman}
\title{Kaiser-Guttman Criterion (Kaiser Line)}
\usage{
kaiser_guttman(
  X = NULL,
  Y = NULL,
  ttmethod = NULL,
  eigs = NULL,
  tol = .Machine$double.eps
)
}
\arguments{
\item{X}{(default: \code{NULL}) first data matrix}

\item{Y}{(default: \code{NULL}) second data matrix}

\item{ttmethod}{(default: \code{NULL}) string denoting the two-table method to use,
either 'plsc', 'cca', or 'rda'.}

\item{eigs}{(default: \code{NULL}) a vector of eigenvalues for the test.}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}
}
\value{
list containing:
\itemize{
\item \code{l}: vector of eigenvalues
\item \code{k_line}: the average of the eigenvalues, above this value is considered
important or significant.
\item \code{imp_comp}: boolean vector saying whether each component is above the
\code{k_line} or not.
}
}
\description{
A simple stopping rule to keep components with an eigenvalue greater than the average
eigenvalue. This average eigenvalue is the Kaiser line or \code{k_line}. The function
can either take \code{X}, \code{Y}, and \code{ttmethod} and compute the results, or a vector
of already computed eigenvalues \code{eigs} can be passed in. This stopping rule is
not recommended.
}
