% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LEMON}
\alias{LEMON}
\title{LEMON Cognitive, Medical, and Demographics}
\format{
contains 3 data.frames with data on the same participants.
The first is the \code{cognitive} data.frame containing the following variables:
\describe{
\item{ID}{The participant's unique identifier}
\item{CVLT_2}{The California Verbal Learning Task (CVLT) assesses verbal
learning and memory capacity and consists of learning and recalling word lists.
This variable is correct recalls after hearing the list for the first time.}
\item{CVLT_3}{Correct recalls after hearing the list for the fifth time.}
\item{CVLT_4}{Proactive interference; score reversed}
\item{CVLT_5}{Retroactive interference; score reversed}
\item{CVLT_6}{Sum of all correct recalls}
\item{CVLT_13}{Delayed recognition memory performance}
\item{CVLT_14}{Overall repetitions; score reversed}
\item{CVLT_15}{Overall intrusions; score reversed}
\item{TMT_1}{The Trail Making Test (TMT) assesses cognitive flexibility by
asking participants to quickly connect circles that are randomly distributed
on a piece of paper. This variable is subtest A time to connect numbers; score
reversed}
\item{TMT_3}{Subtest A number of errors; score reversed}
\item{TMT_5}{Subtest B time to connect numbers; score reversed}
\item{TMT_7}{Subtest B number of errors; score reversed}
\item{WST_1}{The Vocabulary Test (WST; translated from the German) assesses
verbal intelligence level and language comprehension in order to measure
crystallized intelligence. This is the only variable included, which is the
number of words correctly identified on the task.}
\item{LPS_1}{The Performance Testing System (LPS; translated from the German)
assesses logical and inferential thinking in order to measure fluid intelligence.
This is the only variable included, which is the number of correct responses.}
\item{TAP_A_6}{The Test of Attentional Performance (TAP) measures different
aspects of attentional processing using three subtests: Alertness (TAP_A),
Incompatibility (TAP_I), and Working Memory (TAP_WM). This variable is median
reaction time without signal; score reversed}
\item{TAP_A_11}{Median reaction time with signal; score reversed}
\item{TAP_A_15}{Phasic alertness}
\item{TAP_I_2}{Median reaction time for compatible stimuli; score reversed}
\item{TAP_I_6}{Number of errors for compatible stimuli; score reversed}
\item{TAP_I_9}{Median reaction time for incompatible stimuli; score reversed}
\item{TAP_I_13}{Number of errors for incompatible stimuli; score reversed}
\item{TAP_WM_2}{Median reaction time; score reversed}
\item{TAP_WM_6}{Number of correct matches}
\item{TAP_WM_7}{Number of incorrect matches or errors; score reversed}
\item{RWT_8}{The Regensburger Word Fluency Test (RWT) assesses verbal fluency.
This variable is number of S-Words named}
\item{RWT_10}{Number of S-Word repetitions; score reversed}
\item{RWT_11}{Number of S-Word rule breaks; score reversed}
\item{RWT_20}{Number of animal words named}
\item{RWT_22}{Number of animal word repetitions; score reversed}
\item{RWT_23}{Number of animal word rule breaks; score reversed}
}
The second is the \code{medical} data.frame containing the following variables:
\describe{
\item{ID}{The participant's unique identifier}
\item{Height_cm}{Height in centimeters}
\item{Weight_kg}{Body weight in kilograms}
\item{Waist_cm}{Waist circumference in centimeters}
\item{Hip_cm}{Hip circumference in centimeters}
\item{BP1_left_systole}{Systolic blood pressure at time 1}
\item{BP1_left_diastole}{Diastolic blood pressure at time 1}
\item{pulse1_left}{Pulse at time 1}
\item{BP2_left_systole}{Systolic blood pressure at time 2}
\item{BP2_left_diastole}{Diastolic blood pressure at time 2}
\item{pulse2_left}{Pulse at time 2}
\item{BP3_left_systole}{Systolic blood pressure at time 3}
\item{BP3_left_diastole}{Diastolic blood pressure at time 3}
\item{ALAT}{Alanine Aminotransferase}
\item{ASAT}{Aspartate Aminotransferase}
\item{HCT}{Hematocrit}
\item{HGB}{Hemoglobin}
\item{MCH}{Mean Corpuscular Haemoglobin}
\item{MCHC}{Mean Corpuscular Haemoglobin Concentration}
\item{MCV}{Mean Corpuscular Volume}
\item{RBC}{Red Blood Cells}
\item{RDW}{Red Cell Distribution Width}
\item{WBC}{White Blood Cells}
\item{CHOL}{Cholesterol}
\item{CL}{Chloride}
\item{CRE}{Creatinine}
\item{GGT}{Gamma Glutamyl Transferase}
\item{GLU}{Glucose}
\item{HB}{Hemoglobin for Hemoglobin A1c}
\item{HBA1G}{Hemoglobin A1c amount}
\item{HDLC}{HDL-Cholesterol}
\item{K}{Potassium}
\item{LDLC}{LDL-Cholesterol}
\item{NA}{Sodium}
\item{TRIG}{Triglycerides}
\item{TSH}{Thyroid-Stimulating Hormone}
}
The third is the \code{demo} data.frame with the following variables:
\describe{
\item{ID}{The participant's unique identifier}
\item{Gender}{Participant gender, either Male or Female}
\item{Age_bin}{Participant age group, either Young or Mature}
}
}
\source{
https://ftp.gwdg.de/pub/misc/MPI-Leipzig_Mind-Brain-Body-LEMON/
}
\usage{
LEMON
}
\description{
A subset of the Leipzig Study for Mind-Body-Emotion Interactions (LEMON) data,
a publicly available dataset of psychological, physiological, and
neuroimaging measures collected on a sample of 227 participants.
This data is a sample of 144 participants who had complete data on a subset
of cognitive tasks and medical variables.
}
\keyword{datasets}
