% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_mult}
\alias{mat_mult}
\title{Fast matrix multiplication using C++}
\usage{
mat_mult(A, B)
}
\arguments{
\item{A}{left matrix in the multiplication}

\item{B}{right matrix in the multiplication}
}
\value{
a matrix of size \code{nrow(A)} by \code{ncol(B)}
}
\description{
This function offers a faster matrix multiplication than R provides natively.
It takes two matrices, \code{A} and \code{B} and returns the matrix multiplication of
\code{A \%*\% B}.
}
