% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_mult_order.R
\name{matrix_chain_multiplication}
\alias{matrix_chain_multiplication}
\title{Matrix Chain Multiplication}
\usage{
matrix_chain_multiplication(matrices, split = NULL)
}
\arguments{
\item{matrices}{list of compatible matrices to multiply. Matrices are multiplied
in the order they are given in the list.}

\item{split}{(default: NULL) the split matrix computed by \code{get_split_mat} can
be supplied here, or it is computed automatically if \code{split = NULL}.}
}
\value{
a matrix that is the result of multiplying the matrices in \code{matrices}
}
\description{
Given a list of compatible matrices to multiply, this function returns their
multiplication, using an efficient order of matrix multiplications. The order
of matrix multiplications is computed using a matrix chain multiplication
algorithm (see for instance: https://en.wikipedia.org/wiki/Matrix_chain_multiplication).
}
