% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cormats_and_data.R
\name{multi_block_cor}
\alias{multi_block_cor}
\title{Create a correlation matrix between X and Y with multiple blocks of variables}
\usage{
multi_block_cor(
  cor_blocks,
  ncolX,
  ncolY,
  ncolX_blocks,
  ncolY_blocks,
  cor_blocksX,
  cor_blocksY,
  offXY = 0,
  offX = 0,
  offY = 0
)
}
\arguments{
\item{cor_blocks}{a vector of correlation values corresponding to each block.
\code{length(cor_blocks)} determines the number of blocks (\code{n_blocks}).}

\item{ncolX}{integer number of columns that belong to the X matrix}

\item{ncolY}{integer number of columns that belong to the Y matrix}

\item{ncolX_blocks}{vector same length as \code{cor_blocks} that gives the
number of X columns in each block.}

\item{ncolY_blocks}{vector same length as \code{cor_blocks} that gives the
number of Y columns in each block.}

\item{cor_blocksX}{vector the same length as \code{cor_blocks} that contains the
within X correlations for each block}

\item{cor_blocksY}{vector the same length as \code{cor_blocks} that contains the
within X correlations for each block}

\item{offXY}{scalar correlation between variables of X and Y that aren't in a block}

\item{offX}{scalar correlation between variables of X that aren't in a block}

\item{offY}{scalar correlation between variables of Y that aren't in a block}
}
\value{
a correlation matrix based on the parameters, with size
\code{ncolX + ncolY} by \code{ncolX + ncolY}. The first \code{ncolX} rows/columns
correspond to the X variables, and the last \code{ncolY} rows/columns to the
Y variables.
}
\description{
This function creates a joint correlation matrix for the variables of X and Y,
with the main feature being the blocks of correlated variables placed on the
diagonal.
}
