% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_methods.R
\name{plsc}
\alias{plsc}
\title{Partial Least Squares Correlation}
\usage{
plsc(X, Y, tol = .Machine$double.eps)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}
}
\value{
a list of results including:
\itemize{
\item \code{d}: vector of singular values
\item \code{u}: matrix of left singular vectors
\item \code{v}: matrix of right singular vectors
\item \code{l}: vector of eigenvalues
\item \code{lx}: matrix of latent variables for \code{X}
\item \code{ly}: matrix of latent variables for \code{Y}
\item \code{p}: matrix of left generalized singular vectors. For PLSC, \code{p} and \code{u} are identical.
\item \code{q}: matrix of right generalized singular vectors. For PLSC, \code{q} and \code{v} are identical.
\item \code{m}: left metric matrix for the GSVD. For PLSC it is the identity matrix.
\item \code{w}: right metric matrix for the GSVD. For PLSC it is the identity matrix.
\item \code{f}: matrix of loadings for \code{X}. \code{lx} is computed by multiplying \code{X} and \code{f}. For PLSC,
\code{f} and \code{p} are identical.
\item \code{g}: matrix of loadings for \code{Y}. \code{ly} is computed by multiplying \code{Y} and \code{g}. For PLSC,
\code{g} and \code{q} are identical.
}
}
\description{
A basic function for computing results of Partial Least Squares Correlation (PLSC).
Relies on the underlying \code{\link[GSVD:gplssvd]{GSVD::gplssvd()}} method.
}
