% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_POV.R
\name{pov}
\alias{pov}
\title{Percentage of variance (POV) explained}
\usage{
pov(
  X = NULL,
  Y = NULL,
  ttmethod = NULL,
  eigs = NULL,
  tol = .Machine$double.eps
)
}
\arguments{
\item{X}{(default: \code{NULL}) first data matrix}

\item{Y}{(default: \code{NULL}) second data matrix}

\item{ttmethod}{(default: \code{NULL}) string denoting the two-table method to use,
either 'plsc', 'cca', or 'rda'.}

\item{eigs}{(default: \code{NULL}) a vector of eigenvalues for the test.}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}
}
\value{
list containing:
\itemize{
\item \code{l}: the vector of eigenvalues from the analysis
\item \code{inertia}: the total inertia (variance), which is the sum of the eigenvalues
\item \code{perc_inertia}: a vector the same length as \code{l} giving the percentage of
inertia explained by each component.
\item \code{cum_perc_inertia}: a vector the same length as \code{l} giving the cumulative
percentage of inertia explained from the first component up to that component
}
}
\description{
A simple stopping rule which returns the cumulative percentage of variance
or 'inertia' explained by the components of a two-table analysis. One can then
use the cumulative percentage to keep the first set of components which reach
a certain threshold, such as 80\%. The function can either take \code{X}, \code{Y}, and
\code{ttmethod} and compute the results, or a vector of already computed
eigenvalues \code{eigs} can be passed in. This stopping rule is not recommended.
}
