% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{pseudo_R2}
\alias{pseudo_R2}
\title{Compute pseudo-\eqn{R^2} values based on eigenvalues}
\usage{
pseudo_R2(eigs, method = "l+1")
}
\arguments{
\item{eigs}{a vector of eigenvalues for the computation}

\item{method}{"l+1" (default) corresponds to pseudo-\eqn{R^2} 1, and uses
all eigenvalues after the "current" eigenvalue as the divisor of the ratio.
"l" corresponds to pseudo-\eqn{R^2} 2, and uses all eigenvalues after the
"current" eigenvalue plus the current eigenvalue as the divisor of the ratio.}
}
\value{
a vector the same length as \code{eigs} containing the pseudo-\eqn{R^2}s
}
\description{
Given a vector of eigenvalues, this function computes the ratio between each
eigenvalue and the sum of the eigenvalues after it. There are two types of
pseudo-\eqn{R^2}s included: pseudo-\eqn{R^2} 1 includes only later eigenvalues
in the sum, and pseudo-\eqn{R^2} 2 also includes the present eigenvalue in the
sum.
}
