% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{pval_from_dist}
\alias{pval_from_dist}
\title{Compute p value given a statistic and a null distribution for the statistic}
\usage{
pval_from_dist(stat, stat_dist, test_type = "one_sided_right", tol = 1e-06)
}
\arguments{
\item{stat}{a scalar or vector of observed statistics to test}

\item{stat_dist}{a matrix of size \code{length(stat)} by \code{n_iter} (number of permutations).
Also accepts \code{n_iter} by \code{length(stat)}. For each value in \code{stat} there
should be a corresponding row/column of \code{stat_dist} containing its null distribution.
or \code{length(stat)} by \code{n_iter}.}

\item{test_type}{either "one_sided_right" (default), "one_sided_left", or "two_sided".
This is the type of hypothesis test to do, aka one-tailed positive, one-tailed negative,
or two-tailed. Essentially all two-table stopping rules use \code{one_sided_right}.}

\item{tol}{(default: 1e-6) tolerance used for comparisons between \code{stat} and \code{stat_dist}}
}
\value{
a scalar or vector the same length as \code{stat} containing p values for \code{stat}
}
\description{
This function computes p values based on null distributions from
permutation tests. It was written to be used with the results of
\code{\link[=SRP]{SRP()}}, \code{\link[=DRP]{DRP()}}, or \code{\link[=split_half]{split_half()}}. This function should not be used with results
from \code{\link[=multiple_hold]{multiple_hold()}}; see \code{\link[=pvals_for_MH]{pvals_for_MH()}} instead.
}
