% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{pvals_all}
\alias{pvals_all}
\title{Return p values for all statistics and null distributions in a list}
\usage{
pvals_all(dist_list, test_type = "one_sided_right", tol = 1e-06)
}
\arguments{
\item{dist_list}{list containing scalars/vectors that are statistics and
matrices of null distributions. Every statistic must have an analagously named
null distribution, so if there is a vector named "stat" there must be a matrix
named "stat_dist" also in the list.}

\item{test_type}{either "one_sided_right" (default), "one_sided_left", or "two_sided".
This is the type of hypothesis test to do, aka one-tailed positive, one-tailed negative,
or two-tailed. Essentially all two-table stopping rules use \code{one_sided_right}.}

\item{tol}{(default: 1e-6) tolerance used for comparisons between \code{stat} and \code{stat_dist}}
}
\value{
a list of length \code{length(dist_list)/2} containing p values for each
statistic/distribution in \code{dist_list}
}
\description{
\code{pvals_all()} is a wrapper around \code{\link[=pval_from_dist]{pval_from_dist()}} that allows you to easily get
p values for several statistics and null distributions in a list, such as the
\code{perm_res} output of \code{\link[=SRP]{SRP()}}, \code{\link[=DRP]{DRP()}}, or \code{\link[=split_half]{split_half()}}.
}
