% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{pvals_for_MH}
\alias{pvals_for_MH}
\title{Compute p values for the multiple holdout (MH) stopping rule}
\usage{
pvals_for_MH(dist_list, tol = 1e-06)
}
\arguments{
\item{dist_list}{the \code{perm_res} returned by \code{\link[=multiple_hold]{multiple_hold()}} containing
\code{rho} and \code{rho_dist}.}

\item{tol}{(default: 1e-6) tolerance for comparison between \code{rho} and \code{rho_dist}}
}
\value{
list containing:
\itemize{
\item \code{pvals_omni}: a matrix of size \code{n_holdouts} by \code{length(rho)}.
\item \code{pvals_avg}: a vector the same length as \code{rho}
}
}
\description{
This function takes the \code{perm_res} from the results of \code{\link[=multiple_hold]{multiple_hold()}}
and uses it to compute p values in two different ways: the omnibus method
and the averaging method.
}
