% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rand_each_col_cpp}
\alias{rand_each_col_cpp}
\title{Randomize (permute) each column of a matrix}
\usage{
rand_each_col_cpp(X)
}
\arguments{
\item{X}{the matrix to permute}
}
\value{
a matrix with the values in each column of \code{X} reordered
}
\description{
The C++ implementation gives this function more speed than the \code{rand_each_col()}
function written in just in R.
This function randomizes or permutes a matrix by reordering the values in each
column of the matrix. This results in a matrix where the relationships (correlations)
within the matrix are broken, and when compared to another matrix with different
variables and the same participants, the relationships between the two matrices
are also now broken.
}
\details{
This function is the option used when selecting "rand_each_col" in the stopping rule
functions such as \code{\link[=SRP]{SRP()}} or \code{\link[=DRP]{DRP()}}.
}
