% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_mat.R
\name{rand_rows}
\alias{rand_rows}
\title{Randomize (permute) the rows of a matrix}
\usage{
rand_rows(X)
}
\arguments{
\item{X}{the matrix to permute}
}
\value{
a matrix with the rows of \code{X} reordered
}
\description{
This function randomizes or permutes a matrix by reordering complete rows of
the matrix. This results in a matrix where the relationships (correlations)
within the matrix are preserved. But when compared to another matrix with different
variables and the same participants, the relationships between the two matrices
are now broken.
}
\details{
This function is the option used when selecting "rand_rows" in the stopping rule
functions such as \code{\link[=SRP]{SRP()}} or \code{\link[=DRP]{DRP()}}. However, all the stopping rule functions
in this package use a faster version written with \code{Rcpp}.
}
