% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_methods.R
\name{rda}
\alias{rda}
\title{Redundancy Analysis}
\usage{
rda(X, Y, tol = .Machine$double.eps)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}
}
\value{
a list of results including
\itemize{
\item \code{d}: vector of singular values
\item \code{u}: matrix of left singular vectors
\item \code{v}: matrix of right singular vectors
\item \code{l}: vector of eigenvalues
\item \code{lx}: matrix of latent variables for X
\item \code{ly}: matrix of latent variables for Y
\item \code{p}: matrix of left generalized singular vectors
\item \code{q}: matrix of right generalized singular vectors
\item \code{m}: left metric matrix for the GSVD. For RDA, it is the inverse crossproduct of \code{X}.
\item \code{w}: right metric matrix for the GSVD. For RDA, it is the identity matrix.
\item \code{f}: matrix of loadings for \code{X}. \code{lx} is computed by multiplying \code{X} and \code{f}. For RDA,
\code{f} is \code{m} multiplied by \code{p}.
\item \code{g}: matrix of loadings for \code{Y}. \code{ly} is computed by multiplying \code{Y} and \code{g}. For RDA,
\code{g} and \code{q} are identical.
}
}
\description{
A basic function for computing results of Redundancy Analysis (RDA).
Relies on the underlying \code{\link[GSVD:gplssvd]{GSVD::gplssvd()}} method.
}
