% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{rvdim}
\alias{rvdim}
\title{Compute RVDIM or RLSDIM from Dray (2008).}
\usage{
rvdim(eigs, method = "rvdim")
}
\arguments{
\item{eigs}{a vector of eigenvalues for the computation}

\item{method}{"rvdim" (default) or "rlsdim"}
}
\value{
a vector the same length as \code{eigs}
}
\description{
This function computes for each eigenvalue an RVDIM value or an RLSDIM value
as described in Dray (2008).
}
