% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_SS1}
\alias{scale_SS1}
\title{Center and normalize the columns of a matrix to a sum of squares of 1.}
\usage{
scale_SS1(X)
}
\arguments{
\item{X}{matrix to center and scale}
}
\value{
centered and scaled X
}
\description{
Centers each column and scales each column to have a sum of squares of 1.
}
