% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_methods.R
\name{select_tt}
\alias{select_tt}
\title{Select and run a two-table analysis: PLSC, CCA, or RDA}
\usage{
select_tt(X, Y, ttmethod, tol = .Machine$double.eps)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{ttmethod}{string denoting the two-table method to use,
either 'plsc', 'cca', or 'rda'.}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}
}
\value{
results of \code{\link[=plsc]{plsc()}}, \code{\link[=cca]{cca()}}, or \code{\link[=rda]{rda()}}.
}
\description{
This is a wrapper around \code{\link[=plsc]{plsc()}}, \code{\link[=cca]{cca()}}, or \code{\link[=rda]{rda()}}.
The specific two-table analysis to run can be selected using \code{ttmethod}.
}
