% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRP_DRP.R
\name{SRP}
\alias{SRP}
\title{Simultaneous randomization procedure (SRP)}
\usage{
SRP(
  X,
  Y,
  ttmethod,
  n_iter = 99,
  rand_type = "rand_rows",
  tol = .Machine$double.eps,
  parallel = FALSE
)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{ttmethod}{string denoting the two-table method to use,
either 'plsc', 'cca', or 'rda'.}

\item{n_iter}{(default: 99) number of permutation iterations}

\item{rand_type}{(default: 'rand_rows') how to randomize the matrices,
either 'rand_rows'/'RR' or 'rand_each_col'/'REC'.
The RR option permutes entire rows of the \code{X} matrix while leaving \code{Y} unchanged.
The REC option permutes every column of both \code{X} and \code{Y}.}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}

\item{parallel}{(default: FALSE) use a parallelized loop to do the main permutations.
This will use the \code{parallel} package to utilize multiple cores simultaneously
(using \code{\link[parallel:detectCores]{parallel::detectCores()}} - 1 cores).}
}
\value{
a list containing:
\itemize{
\item \code{fixed_res}: a list containing results from the original, unpermuted analysis
\itemize{
\item \code{d}: singular values from the two-table analysis
\item \code{l}: eigenvalues from the two-table analysis
}
\item \code{perm_res}: a list containing results from the permutation iterations.
For each test statistic, there is a vector of the observed test statistics
for each component, and there is a matrix of the null distribution for each
component obtained from the permutation iterations.
\itemize{
\item \code{d}: a vector of singular values
\item \code{d_dist}: a matrix of size \code{n_iter + 1} by \code{length(d)} containing the
null distributions of the singular values
\item \code{l_prop}: a vector of the eigenvalue proportions (EVP). EVP is computed as
a component's eigenvalue divided by the sum of all the eigenvalues.
\item \code{l_prop_dist}: a matrix of size \code{n_iter + 1} by \code{length(l_prop)} containing the
null distributions of the EVPs.
\item \code{pseudo_R2_1}: a vector of pseudo-\eqn{R^2} 1 values. Pseudo-\eqn{R^2} 1 is
computed as a component's eigenvalue divided by the sum of all the following
eigenvalues, in contrast to pseudo-\eqn{R^2} 2 below.
The last component is assigned \code{NA}.
\item \code{pseudo_R2_1_dist}: a matrix of size \code{n_iter + 1} by \code{length(pseudo_R2_1)}
containing the null distributions of the pseudo-\eqn{R^2} 1 values. The last
column corresponding to the last component are all \code{NA}.
\item \code{pseudo_R2_2}: a vector of pseudo-\eqn{R^2} 2 values. Pseudo-\eqn{R^2} 2 is
computed as a component's eigenvalue divided by the sum of itself and all the following
eigenvalues, in contrast to pseudo-\eqn{R^2} 1 above.
The last component is assigned \code{NA}.
\item \code{pseudo_R2_2_dist}: a matrix of size \code{n_iter + 1} by \code{length(pseudo_R2_2)}
containing the null distributions of the pseudo-\eqn{R^2} 2 values. The last
column corresponding to the last component are all \code{NA}.
\item \code{l_diff}: a vector containing the eigenvalue differences, the difference between
a component's eigenvalue and the following eigenvalue. The last component is assigned \code{NA}.
\item \code{l_diff_dist}: a matrix of size \code{n_iter + 1} by \code{length(l_diff)} containing the
null distributions of the eigenvalue differences. The last
column corresponding to the last component are all \code{NA}.
\item \code{l_ratio}: a vector containing the eigenvalue ratios, the ratio between
a component's eigenvalue and the following eigenvalue. The last component is assigned \code{NA}.
\item \code{l_ratio_dist}: a matrix of size \code{n_iter + 1} by \code{length(l_ratio)} containing the
null distributions of the eigenvalue ratios. The last
column corresponding to the last component are all \code{NA}.
\item \code{rvdim}: a vector containing the RVDIM values, a value proposed by Dray (2008) that
assesses the similarity between the rank-one matrix created by a component, and
the matrix created by all of the following components. The last component is assigned \code{NA}.
\item \code{rvdim_dist}: a matrix of size \code{n_iter + 1} by \code{length(rvdim)} containing the
null distributions of the RVDIM values. The last
column corresponding to the last component are all \code{NA}.
}
}
}
\description{
SRP is a two-table stopping rule that runs a simple permutation test without deflation.
The function computes several test statistics and their null distributions.
It does not actually compute the p values for each component; that is accomplished by
passing the results to \code{\link[=pval_from_dist]{pval_from_dist()}} or \code{\link[=pvals_all]{pvals_all()}}.
}
