% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_methods.R
\name{tt_fast_dl}
\alias{tt_fast_dl}
\title{Fast singular values and eigenvalues for two-table methods}
\usage{
tt_fast_dl(
  X,
  Y,
  ttmethod,
  tol = .Machine$double.eps,
  Rx_invsqrt = NULL,
  Ry_invsqrt = NULL
)
}
\arguments{
\item{X}{first data matrix}

\item{Y}{second data matrix}

\item{ttmethod}{string denoting the two-table method to use,
either 'plsc', 'cca', or 'rda'.}

\item{tol}{(default: \code{.Machine$double.eps}) tolerance value for eigenvalues}

\item{Rx_invsqrt}{NULL (default) or
a precomputed Rx inverse sqrt matrix (if you want to go faster)}

\item{Ry_invsqrt}{NULL (default) or
a precomputed Ry inverse sqrt matrix (if you want to go faster)}
}
\value{
a list with singular values (\code{d}) and eigenvalues (\code{l})
}
\description{
Rather than compute all parts of the two-table methods,
\code{tt_fast_dl()} only computes the the singular values and eigenvalues.
Helpful for permutation tests when the goal is to get these values
with minimum computations.
}
